So I decided to do something about scrolling
I've had this issue with how the message box can expand over text. Let's
say you're replying to someone, but you're looking back up at their
message from time to time for your answer. If your reply covers two
lines, the text box expands over the last message, obscuring the text -
you have to scroll down to see it again.

I decided to do something about it. So now, when it expands the text
box, it won't expand it over the message at the bottom - it'll
automatically reset its scroll position to the bottom.

Except I immediately saw a flaw in this. If you were reading something
not at the bottom of the message log, autoscrolling to the bottom
wouldn't help you. This feature disables itself if you are not at the
bottom of the message log to prevent issues.