4AM Error handling
No more emails because someone left their messenger on.
Now we have a catch for invalid cast exceptions.
This means if the arraylist is in fact a string, we don't drop
everything.

It also resets the notifications boolean so we'll continue to get
notifications when we actually get them.

An issue I saw happening was if somehow, something else triggers the
exception, like something in the array was changed to no longer be
numeric when it looks for a number. If that happens, it'll keep firing
the same notification over and over. With that in mind, I seperated the
setup of the array list and the setup of notifications, barring any
notifications from appearing if we dont get any.

As an added precaution, in the event the server goes ABSOLUTELY NUTS in
the middle of the day, we can tell messenger to quit though the error
message it'll give. Clicking Yes sends an error report, No does nothing,
and Cancel closes the program.